+++ Makefile
@@ -1,5 +1,6 @@
 MAIN = srcs/main/prompt.c \
 		srcs/parsing/lexer.c \
+		srcs/parsing/list_helpers.c
 
 CC = cc
 CFLAGS = -Wall -Wextra -Werror
