+++ includes/parsing.h
@@ -6,7 +6,7 @@
 /*   By: jbarbay <jbarbay@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/15 22:28:38 by jbarbay           #+#    #+#             */
-/*   Updated: 2024/01/18 16:22:32 by jbarbay          ###   ########.fr       */
+/*   Updated: 2024/01/18 18:42:34 by jbarbay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,6 +37,8 @@ int		process_quote(char **line, int quote, t_token **list);
 int		process_pipe(t_token **list);
 int		process_variable(char **line, t_token **list);
 int		process_input(char **line, char *c, t_token **list);
+int		error_tokens(int error, t_token **list);
+
 
 // Token list helpers
 void	print_tokens(t_token *token);
