+++ srcs/parsing/lexer.c
@@ -6,7 +6,7 @@
 /*   By: jbarbay <jbarbay@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/17 12:32:45 by jbarbay           #+#    #+#             */
-/*   Updated: 2024/01/18 16:23:42 by jbarbay          ###   ########.fr       */
+/*   Updated: 2024/01/18 19:01:11 by jbarbay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,11 +19,7 @@ int	process_quote(char **line, int quote, t_token **list)
 
 	i = 1;
 	if (!ft_strchr(*line + 1, quote))
-	{
-		printf("Syntax error: unclosed quotes\n");
-		free_tokens(list);
-		return (-1);
-	}
+		error_tokens(3, list);
 	while ((*line + i) && *(*line + i) != quote)
 		i++;
 	i++;
@@ -37,26 +33,30 @@ int	process_quote(char **line, int quote, t_token **list)
 	return (i);
 }
 
+int	error_tokens(int error, t_token **list)
+{
+	if (error == 1)
+		printf("Environment variable cannot start with a digit\n");
+	else if (error == 2)
+		printf("Only uppercase, lowercase, and underscores are allowed for environment variables\n");
+	else if (error == 3)
+		printf("Syntax error: unclosed quotes\n");
+	free_tokens(list);
+	return (-1);
+}
+
 int	process_variable(char **line, t_token **list)
 {
-	int	i;
+	int		i;
 	char	*str;
 
 	i = 1;
 	if (ft_isdigit((*line)[i]))
+		return (error_tokens(1, list));
+	while (*line + i && !ft_strchr(WHITESPACE, (*line)[i]))
 	{
-		printf("Environment variable cannot start with a digit\n");
-		free_tokens(list);
-		return (-1);
-	}
-	while ((*line)[i] && !ft_strchr(WHITESPACE, (*line)[i]))
-	{	
 		if (!ft_isalnum((*line)[i]) && (*line)[i] != '_')
-		{
-			printf("Only uppercase, lowercase, and underscores are allowed for environment variables\n");
-			free_tokens(list);
-			return (-1);
-		}
+			return (error_tokens(2, list));
 		i++;
 	}
 	str = (char *)malloc(sizeof(char) * (i + 1));
@@ -78,6 +78,7 @@ int	process_pipe(t_token **list)
 int	process_input(char **line, char *c, t_token **list)
 {
 	char	*str;
+
 	if (ft_strncmp((*line + 1), c, 1) == 0)
 	{
 		str = ft_strjoin((char *)c, (char *)c);
@@ -98,9 +99,82 @@ int	process_input(char **line, char *c, t_token **list)
 	}
 }
 
+// We can have quotes inside string
+// Example echo"hello""hello" -> cmd not found echohellohello
+// Make sure the quotes are closed
+// pwd'd' -> valid, pw'd -> not valid
+// echo"hello'hi" -> valid as a token, echo"hello -> invalid token
+
+int	find_quote(const char *s, int c)
+{
+	int	i;
+
+	i = 0;
+	while (s[i])
+	{
+		if (s[i] == (unsigned char)c)
+			return (i);
+		i++;
+	}
+	return (0);
+}
+
+int	check_unclosed_quotes(char *str, t_token **list)
+{
+	int	i;
+	int	pos;
+
+	i = 1;
+	while (str[i])
+	{
+		//  If we find the quote, we have to find the second one
+		if (str[i] == 34)
+		{
+			pos = find_quote(str + i + 1, 34);
+			if (!pos)
+				return (error_tokens(3, list));
+			else
+				i += pos;
+		}
+		else if (str[i] == 39)
+		{
+			pos = find_quote(str + i + 1, 39);
+			if (!pos)
+				return (error_tokens(3, list));
+			else
+				i += pos;
+		}
+		else
+			i++;
+	}
+	return (i);
+}
+
+// Anything else will be STRING
+// cmd1|cmd2>outfile will work, recognizes the pipe and redirections
+// Something like $VAR could be stored as STRING, so some values inside STRING might need to get expanded
+// Example: p$VAR -> during execution expanded as pwd
+
+int	process_string(char **line, t_token **list)
+{
+	int		i;
+	char	*str;
+
+	i = 0;
+	while ((*line)[i] && (*line)[i] != '|' && (*line)[i] != '<' && (*line)[i] != '>')
+		i++;
+	str = (char *)malloc(sizeof(char) * (i + 1));
+	ft_strlcpy(str, *line, i + 1);
+	create_token(STRING, str, list);
+	if (check_unclosed_quotes(str, list) == -1)
+		return (-1);
+	*line += i;
+	return (0);
+}
+
 t_token	*get_tokens(char *line)
 {
-	t_token *list;
+	t_token	*list;
 
 	list = NULL;
 	while (*line)
@@ -129,69 +203,8 @@ t_token	*get_tokens(char *line)
 		else if (*line == '|')
 			line += process_pipe(&list);
 		else
-			line++;
+			if (process_string(&line, &list) == -1)
+				return (NULL);
 	}
 	return (list);
 }
-
-void	create_token(int type, char *str, t_token **list)
-{
-	t_token	*new;
-
-	new = (t_token *)malloc(sizeof(t_token));
-	// if (!new)
-	// 	return (NULL);
-	new->type = type;
-	new->value = str;
-	new->next = NULL;
-	ft_token_add_back(list, new);
-}
-
-void	ft_token_add_back(t_token **lst, t_token *new)
-{
-	t_token	*temp;
-
-	if (*lst == NULL)
-	{
-		*lst = new;
-		return ;
-	}
-	temp = *lst;
-	while (temp->next)
-	{
-		temp = temp->next;
-	}
-	temp->next = new;
-}
-
-void	print_tokens(t_token *token)
-{
-	int	i;
-
-	i = 0;
-	printf("Printing TOKENS: \n");
-	while (token)
-	{
-		printf("Item %d:, type: %d, value: %s\n", i, token->type, token->value);
-		i++;
-		token = token->next;
-	}
-}
-
-void	free_tokens(t_token **list)
-{
-	t_token	*current;
-	t_token	*next;
-
-	current = *list;
-	while (current)
-	{
-		next = current->next;
-		free(current->value);
-		current->value = NULL;
-		free(current);
-		current = next;
-	}
-	*list = NULL;
-}
-
